backend/requirements.txt
streamlit
pymongo[srv]
sentence-transformers
faiss-cpu
langchain-ollama
langchain-core
pandas
matplotlib
wordcloud
tqdm
